requirements.txt
Flask
scikit-learn
pandas
matplotlib
gunicorn 
tweepy
python-dotenv